refactor: removed unused props of AlertMessage component
