+++ src/modules/general/components/AlertMessage/index.types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react';
+
+export interface AlertMessageProps {
+  theme: 'primary' | 'gray' | 'error' | 'warning' | 'success';
+  iconName: string;
+  title: string;
+  subtitle?: string;
+  children?: ReactNode;
+  colOrderMobileView?: boolean;
+  containerClassName?: string;
+}
