+++ src/modules/userProfile/components/profileHeader/index.tsx
@@ -1,7 +1,7 @@
 import { IconButton as MUIIconButton } from '@mui/material';
 import { ConnectRequestModal } from 'src/modules/connections/connectRequestModal';
 import { ThreeDotsButton } from 'src/modules/connections/threeDotsButton';
-import { AlertMessage } from 'src/modules/general/components/alertMessage';
+import AlertMessage from 'src/modules/general/components/AlertMessage';
 import { Button } from 'src/modules/general/components/Button';
 import { Icon } from 'src/modules/general/components/Icon';
 import { EditInfoModal } from 'src/modules/userProfile/containers/editInfo';
